requirements.txt
discord
discord[voice]
ffmpeg
python-dotenv
yt-dlp
async_timeout
validators
openai
